+++ packages/react-native-executorch/src/hooks/natural_language_processing/useLLM.ts
@@ -67,7 +67,9 @@ export const useLLM = ({
     })();
 
     return () => {
-      controllerInstance.delete();
+      if (controllerInstance.isReady) {
+        controllerInstance.delete();
+      }
     };
   }, [
     controllerInstance,
